+++ castle/cms/static/components/upload.js
@@ -211,7 +211,8 @@ define([
       if(this.state.workflow_state !== 'published'){
         stateAction = [
           'This file is currently not published. ',
-          D.a({ href: '#', onClick: this.publishClicked }, 'Publish immediately')
+          D.a({ className: 'content-publish',
+                href: '#', onClick: this.publishClicked }, 'Publish immediately')
         ];
       }
       var label = 'Finished uploading ';
@@ -224,7 +225,9 @@ define([
       return D.div({className: 'finished-container'}, [
         D.p({}, [
           label,
-          D.a({href: this.state.base_url + '/view', target: '_blank'}, this.state.title),
+          D.a({ className: 'content-link',
+                href: this.state.base_url + '/view',
+                target: '_blank'}, this.state.title),
           '. '
         ].concat(stateAction)),
       ]);
@@ -262,7 +265,7 @@ define([
       }
 
       var id = 'castle-upload-field-' + name;
-      return D.div({ className: 'form-group' }, [
+      return D.div({ className: 'form-group upload-field-' + name }, [
         D.label({ className: labelClass, for_: id}, field['label']),
         D.div({className: 'col-sm-8' }, input)
       ]);